v0.7-Supplier-Agents/requirements.txt
mcp[cli]>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
httpx>=0.27.0
fastapi>=0.111.0
uvicorn>=0.29.0